Created architecture diagram
Using the diagrams python library to programmatically create the diagram
picture. I tried to show the technologies we're using and that the main
design pattern is MVC.

Created architecure diagram

Used diagrams python library to create the diagram picture. Tried to
show the technologies we're using and that the main design pattern we're
following is MVC.