LUK-37 created unit tests
Made unit tests for all components. Went off of the sample FetchTest file, and expanded a lot on it.

I couldn't get the test in Courses.test.js to pass, as for some reason the spy function is not properly mocking the api call (or registering that the call is being made). I tried fixing this but couldn't find a good answer anywhere.