LUK-106 added page tracking to website
Added a listener, usePageViews(), to the root component of the frontend that makes a put request whenever a course page is visited.

On the backend we keep track of page views by incrementing a counter.

It is important to remove all forward slashes from page names to avoid routing issues with the REST API