LUK-106 analytics routes and api
Each API is pretty much the same. The idea is to take in the course id, make sure the user making the request is the instructor for the course, then find out whatever information you're looking for.

I made api's for: getting the view count on a course's home page, get the number of enrolled students, and getting the number of published quizzes.